load notes from localStorage to remove refresh delay
When we refresh the page there is a noticeable delay as we need to hit
firebase to get the notes.

Instead, we attempt to load notes from localStorage and convert the
data into an OrderedMap using deepMerge, and then serve that right
up. deepMerge will convert any javascript objects into maps so we had to
change the style property on the note to a Map instead of an object.

We write to the localStorage whenever NoteStorage emits an event.