Fix deleted notes weren't being deleted from local copy
If a note was deleted else where, the localStorage copy will still be
around and there is nothing telling that copy to remove the deleted
note.

To fix this behavior, load localStorage notes into a separate
orderedMap that we call `coldNotesByName` and have the UI use that
data structure on a cold start.

When notes come in from the server, they are added to `_notesByName` and
then have the UI load that instead of cold. Deleted notes should now go
away.