Don't store localHidden in localStorage
We always want to initialize localHidden to the note's hidden property
when we reload the page, so we shouldn't be persisting it. To remove
localHidden from being persisted, we want to remove that property when
serializing Note objects.

Normally, we would just define our own `toJSON` method on `Note`.
However, this won't work with the `Immutable` library since when
serializing a `Map` it will call `toJS` recursively instead of
`toJSON`, so define our own `toJS` method instead. Perhaps this will get
fixed in the future and when it does we really should be overriding
`toJSON` instead of `toJS`.