Make Index.jsx dumber by extracting Firebase calls into API module
The API module is now the only place that knows about the Firebase ref
(besides Appconfig). It can be interfaced using `start()` and `stop()`
which sets up and removes the firebase listeners.

Writes will now be performed on Actions directly talking to the API.
Reads still happen from the Store. Hopefully this will make things
easier to test.