Add NoteStore test
In the process of adding this test, I changed onToggleLocalHidden to
take a note name instead of note object, since it doesn't make sense to
do anything if the note doesn't already exist in the store. In addition,
I used `updateIn` to do a deep change in the notes structure which was
pretty cool.