Add test for `NoteStore::persist`
Some nuances: `gulp-react` complains about using `import` in a
function. I might just revert back to using CJS syntax. Not really
digging import.

The call to require localStorage has to be called in before each. If
called outside, the localStorage in the test will refer to a different
object from the localStorage in the NoteStore implementation.