Fix handling onNoteChanged
We were merging an unserialized note with the note in the store. This
meant the `createdAt` prop on the note could have changed from a `Date`
object into an ISO 8601 string. Moment doesn't care either way, but for
consistency we should just deserialize the note, which means performing
the same thing as `onNoteAdded`. We assume the full representation of
the note from the backend to come in.

Also, test that `triggerAsync` is called on NoteStore `on` methods.