Remove toArray() conversion
React 0.13 accepts iterables now so we don't have to convert the
sequence to an array. We are converting it to an IndexedSeq instead of a
Seq because and IndexedSeq discard keys. React might deprecate support
for maps (key value pairs) so it's better to just use an indexed seq for
now.